Deleteの修正

+++ app/api/item/delete/[id]/route.js
@@ -3,14 +3,24 @@ import connectDB from "@/app/utils/database";
 import { ItemModel } from "@/app/utils/schemaModels";
 
 export async function DELETE(request, context){
+  const reqBody = await request.json();
   try {
     await connectDB()
-    await ItemModel.deleteOne({
-      _id: context.params.id
-    })
-    return NextResponse.json({
-      message: "削除成功"
-    })
+    const singleItem = await ItemModel.findById(context.params.id);
+    if(singleItem.email == reqBody.email){
+      await ItemModel.deleteOne({
+        _id: context.params.id
+      })
+      return NextResponse.json({
+        message: "削除成功"
+      })
+    } else {
+      return NextResponse.json({
+        message: "他の人が作成したアイテムです"
+      })
+    }
+    
+    
   } catch {
     return NextResponse.json({
       message: "削除失敗"
